+++ tests/acceptance/pagination-test.js
@@ -53,7 +53,7 @@ module('Acceptance: Pagination', {
   beforeEach: function() {
     application = startApp();
 
-    store = application.__container__.lookup('store:main');
+    store = application.__container__.lookup('service:store');
 
     // The implementation of the paginated Pretender server is dynamic
     // so it can be used with all of the pagination tests. Otherwise,
@@ -111,9 +111,9 @@ module('Acceptance: Pagination', {
 });
 
 test('Retrieve list of paginated records', function(assert) {
-  assert.expect(8);
+  assert.expect(7);
 
-  return store.findAll('post').then(function(response) {
+  return store.query('post', {page: 1}).then(function(response) {
     assert.ok(response);
 
     assert.equal(response.get('length'), 4);
@@ -123,75 +123,53 @@ test('Retrieve list of paginated records', function(assert) {
     assert.equal(post.get('postTitle'), 'post title 2');
     assert.equal(post.get('body'), 'post body 2');
 
-    // Test the type metadata.
-    var metadata = store.metadataFor('post');
+    var metadata = response.get('meta');
     assert.equal(metadata.count, 6);
     assert.equal(metadata.next, 2);
     assert.equal(metadata.previous, null);
-
-    // No metadata on results when using find without query params.
-    assert.ok(!response.get('meta'));
   });
 });
 
 
 test("Type metadata doesn't have previous", function(assert) {
-  assert.expect(5);
+  assert.expect(4);
 
-  return store.findAll('post').then(function(response) {
+  return store.query('post', {page: 1}).then(function(response) {
     assert.ok(response);
 
-    // Test the type metadata.
-    var metadata = store.metadataFor('post');
+    var metadata = response.get('meta');
     assert.equal(metadata.count, 6);
     assert.equal(metadata.next, 2);
     assert.equal(metadata.previous, null);
-
-    // No metadata on results when using findAll.
-    assert.ok(!response.get('meta'));
   });
 });
 
 
 test("Type metadata doesn't have next", function(assert) {
-  assert.expect(8);
+  assert.expect(5);
 
   return store.query('post', {page: 2}).then(function(response) {
     assert.ok(response);
     assert.equal(response.get('length'), 2);
 
-    // Test the type metadata.
-    var typeMetadata = store.metadataFor('post');
-    assert.equal(typeMetadata.count, 6);
-    assert.equal(typeMetadata.next, null);
-    assert.equal(typeMetadata.previous, 1);
-
-    // Test the results metadata.
-    var resultsMetadata = response.get('meta');
-    assert.equal(resultsMetadata.count, 6);
-    assert.equal(resultsMetadata.next, null);
-    assert.equal(resultsMetadata.previous, 1);
+    var metadata = response.get('meta');
+    assert.equal(metadata.count, 6);
+    assert.equal(metadata.next, null);
+    assert.equal(metadata.previous, 1);
   });
 });
 
 
 test("Test page_size query param", function(assert) {
-  assert.expect(8);
+  assert.expect(5);
 
   return store.query('post', {page: 2, page_size: 2}).then(function(response) {
     assert.ok(response);
     assert.equal(response.get('length'), 2);
 
-    // Test the type metadata.
-    var typeMetadata = store.metadataFor('post');
-    assert.equal(typeMetadata.count, 6);
-    assert.equal(typeMetadata.previous, 1);
-    assert.equal(typeMetadata.next, 3);
-
-    // Test the results metadata.
-    var resultsMetadata = response.get('meta');
-    assert.equal(resultsMetadata.count, 6);
-    assert.equal(resultsMetadata.previous, 1);
-    assert.equal(resultsMetadata.next, 3);
+    var metadata = response.get('meta');
+    assert.equal(metadata.count, 6);
+    assert.equal(metadata.previous, 1);
+    assert.equal(metadata.next, 3);
   });
 });